Conflicts:
	app/helpers/solution_helper.rb

+++ app/helpers/solution/articles_helper.rb
@@ -24,6 +24,7 @@ module Solution::ArticlesHelper
   end
   
   def draft_info_text
+    @current_item ||= @article.draft || @article
     if @article.draft and @article.draft.locked?
       t('solution.articles.restrict_edit', :name => h(@current_item.user.name)).html_safe
     else
@@ -36,6 +37,7 @@ module Solution::ArticlesHelper
   end
   
   def discard_link
+    return unless privilege?(:delete_solution)
     link_to(t('solutions.drafts.discard'), solution_article_draft_path(@article.id, @article.draft), 
               :method => 'delete',
               :confirm => t('solution.articles.draft.discard_confirm'),
@@ -44,7 +46,7 @@ module Solution::ArticlesHelper
   end
 
   def publish_link
-    return if @article.folder.is_default?
+    return if @article.folder.is_default? or !privilege?(:publish_solution)
     link_to(t('solutions.drafts.publish'), publish_solution_draft_path(@article), 
               :method => 'post', 
               :class => 'draft-btn') if (@article.draft.present? || @article.status == Solution::Article::STATUS_KEYS_BY_TOKEN[:draft])
@@ -127,6 +129,15 @@ module Solution::ArticlesHelper
               "submit-label" => t('save'), 
               "submit-loading" => t('saving')
             }).html_safe
+  end
+
+
+  def draft_saved_notif_bar
+    %(<span> #{t('solution.draft.autosave.save_success')} </span>
+      <span title="#{formated_date(@article.draft.updated_at)}" class="tooltip" data-livestamp="#{@article.draft.updation_timestamp}"></span>
+      <span class="pull-right">
+        #{link_to t('solution.articles.view_draft'), support_draft_preview_path(@article, "preview"),:target => "draft-"+@article.id.to_s}
+      </span>).html_safe
   end
   
 end
